Pi-List-Collector
Read a list of URLS that have ad and malware domain name, collect them , clean them, then rip out the TLD blocks from the wilds file, the wilds file is loaded in to the pi hole as wild card blocks, the final.lists.txt file is push to a S3 bucket , the s3 bucket serves a web page that I add to the pihole as a groups ad list. No the prettiest script , but works well, the final.lists.txt is very clean, the TLD blocks are things like "*.ru" , the for loop removes domain names ending in one of the TDL blocks, thus the final.lists.txt is kept as short as is practical. Millage may vary.